Improve emoji height handling
The EmojiSpan is now drawn with the exact line height of the TextView it is in. This avoids jumping lines when typing in an EditText or weird visuals when having a line with and one without emojis in a TextView.

Furthermore the custom attr "emojiSize" was removed, as it is now automatically calculated. To still be able to choose the size, the textAppearance of the TextView should be changed.

Another change is a large textAppearance for the adapter items, to showcase different heights.