Improve the speed of scanning the text for emojis
This is done by using a regex-based approach now. The EmojiTree has been removed and during the EmojiManagager#install method, we put together a regex which is then used in the new EmojiManager#findAllEmojis method.
This also moves the logic for finding the emojis in a String to the EmojiManager. This was done as the EmojiManager seems to be a semantically better location. This also allows us to later make its interface public to use it for other things.